Separates the XML encoding from the HTTP transport project
