+++ bindings/encoding-xml/src/main/java/esa/mo/mal/encoder/http/HTTPXMLElementInputStream.java
