+++ bindings/encoding-xml/src/main/java/esa/mo/mal/encoder/http/HTTPXMLStreamFactory.java
@@ -23,6 +23,7 @@ package esa.mo.mal.encoder.http;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.Map;
+import java.util.logging.Logger;
 import org.ccsds.moims.mo.mal.encoding.MALElementInputStream;
 import org.ccsds.moims.mo.mal.encoding.MALElementOutputStream;
 import org.ccsds.moims.mo.mal.encoding.MALElementStreamFactory;
@@ -32,6 +33,11 @@ import org.ccsds.moims.mo.mal.MALException;
  *
  */
 public class HTTPXMLStreamFactory extends MALElementStreamFactory {
+
+    /**
+     * Logger
+     */
+    public static final Logger RLOGGER = Logger.getLogger("org.ccsds.moims.mo.mal.transport.http");
 
     @Override
     protected void init(String protocol, Map properties)
