+++ bindings/encoding-xml/src/main/java/esa/mo/mal/encoder/http/HTTPXMLStreamListReader.java
@@ -20,48 +20,59 @@
  */
 package esa.mo.mal.encoder.http;
 
-import static esa.mo.mal.transport.http.HTTPTransport.RLOGGER;
+import static esa.mo.mal.encoder.http.HTTPXMLStreamFactory.RLOGGER;
 import java.util.List;
 import java.util.logging.Level;
+import javax.xml.stream.events.XMLEvent;
+import javax.xml.stream.XMLEventReader;
 import javax.xml.stream.XMLStreamException;
-import javax.xml.stream.XMLStreamWriter;
-import org.ccsds.moims.mo.mal.structures.Element;
+import org.ccsds.moims.mo.mal.MALListDecoder;
 
-public class HTTPXMLStreamListWriter extends HTTPXMLStreamWriter {
+public class HTTPXMLStreamListReader implements MALListDecoder {
 
+    private XMLEventReader eventReader;
     private List list;
+    private String listName;
+    private int size = 0;
 
-    public HTTPXMLStreamListWriter(XMLStreamWriter wr, List list) {
-        this(wr, list, list.getClass().getSimpleName());
+    public HTTPXMLStreamListReader(List list, XMLEventReader eventReader) {
+        this(list, eventReader, list.getClass().getSimpleName());
     }
 
-    public HTTPXMLStreamListWriter(XMLStreamWriter wr, List list, String typeName) {
-        this.writer = wr;
+    public HTTPXMLStreamListReader(List list, XMLEventReader eventReader, String elementName) {
         this.list = list;
-
-        try {
-            writer.writeDTD(LINE_END);
-            writer.writeStartElement(typeName);
-            if (!typeName.equals("AttributeList")) {
-                writer.writeAttribute("malxml:type", ((Element) list).getShortForm().toString());
-            }
-            writer.writeDTD(LINE_END);
-        } catch (XMLStreamException e) {
-            RLOGGER.log(Level.SEVERE, e.getMessage(), e);
-        }
-    }
-
-    public int size() {
-        return list.size();
+        this.listName = elementName;
+        this.eventReader = eventReader;
     }
 
     @Override
-    public void close() {
+    public boolean hasNext() {
         try {
-            writer.writeDTD(LINE_END);
-            writer.writeEndElement();
+            if (!eventReader.hasNext()) {
+                return false;
+            }
+
+            XMLEvent event = eventReader.peek();
+
+            if (event.isCharacters()) {
+                eventReader.nextEvent();
+            }
+            boolean isStart = event.isStartElement();
+
+            if (isStart) {
+                this.size++;
+            }
+            return isStart;
         } catch (XMLStreamException e) {
             RLOGGER.log(Level.SEVERE, e.getMessage(), e);
         }
+
+        return false;
     }
+
+    @Override
+    public int size() {
+        return this.size;
+    }
+
 }
