+++ bindings/encoding-xml/src/test/java/esa/mo/mal/encoder/http/test/XMLTestHelper.java
